New language d.s.m.structurecheck for (editor) unit tests
It allows to check the structure of any object tree. This is useful for editor
test to check the editor state.

Example:

list<Object> root =
    new arraylist<Object>{"a", 2, new arraylist<Integer>{10, 20}, 3};

check structure of root {
  sequence thisElement contains unordered allOrMore {
    element x1 of type Integer {
      property 2 = thisElement
    }
    element x1 of type string {
      condition thisElement != null
      property "a" = thisElement
    }
    element x1 of type list<Integer> {
      sequence thisElement contains ordered exactly {
        element x1 of type <no subtype> {
          property 10 = thisElement
        }
        element x1 of type <no subtype> {
          property 20 = thisElement
        }
      }
    }
  }
};

The power is in the sequence (child collections) check.
It allows to ignore the order and/or additional/missing elements.
You can for example check if:
- all of the given elements are somewhere in the sequence
- all of the given elements are in the sequence in the given order but
  ignore additional elements at any position
- all of the given elements are NOT in the sequence
- the sequence contains only elements out of the given set (not all)

At the moment you cannot click on the stack trace of the exception to go to the
rule that failed. That's because the exception is thrown from one code location.
Each rule has to generate it's own throw statement to make this work.