Linear mouse selection (free text selection with partial copy)
Hold the alt key during the mouse selection to select the text
independant from the structure of the editor or model.
It is then possible to copy only the nodes in the selected text.
For example, select the method signature and Cmd+C will copy the
method without the statements in the body.

Is a bit buggy and ugly but works in general.